style(types): move `multiple-responses` to `question` subdir
