+++ client/app/bundles/course/assessment/question/multiple-responses/components/OptionsManager.tsx
@@ -10,7 +10,7 @@ import { DragDropContext, Droppable, DropResult } from 'react-beautiful-dnd';
 import { Add } from '@mui/icons-material';
 import { Button, Paper, Typography } from '@mui/material';
 import produce from 'immer';
-import { OptionEntity } from 'types/course/assessment/multiple-responses';
+import { OptionEntity } from 'types/course/assessment/question/multiple-responses';
 
 import { formatErrorMessage } from 'lib/components/form/fields/utils/mapError';
