+++ client/app/types/course/assessment/question/multiple-responses.ts
@@ -1,4 +1,4 @@
-import { AvailableSkills, QuestionData, QuestionFormData } from './questions';
+import { AvailableSkills, QuestionData, QuestionFormData } from '../questions';
 
 interface OptionListData {
   id: number | string;
